Rename and reorganise pages (again)
