+++ code/beastly-brawl-showdown/imports/ui/HomePage.jsx
@@ -1,29 +1,16 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import "./HomePage.css";
 
-export default function HomePage() {
+export const HomePage = () => {
   const navigate = useNavigate();
-
   return (
-    <div className="homepage-container">
-      <div className="logo-container">Put logo here</div>
-      <div className="buttons-container">
-        <button className="btn host-btn" onClick={() => navigate("/host")}>
-          HOST
-        </button>
-        <button className="btn join-btn" onClick={() => navigate("/join")}>
-          JOIN
-        </button>
-      </div>
-    </div>
-  );
-}
-<>
-<h1>Home</h1>
+    <>
+      <h1>Home</h1>
 
-<div>
-  <button onClick={() => navigate("/host")}>Host Room</button>
-  <button onClick={() => navigate("/join")}>Join Room</button>
-</div>
-</>
+      <div>
+        <button onClick={() => navigate("/host")}>Host Room</button>
+        <button onClick={() => navigate("/join")}>Join Room</button>
+      </div>
+    </>
+  );
+};
