--- code/beastly-brawl-showdown/imports/ui/HostPage.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-import "./HostPage.css";
-
-export default function HostPage() {
-  return (
-    <div className="page-container">
-      <h1>Host Page</h1>
-    </div>
-  );
-}
