--- code/beastly-brawl-showdown/imports/ui/JoinForm.jsx
@@ -1,30 +1,0 @@
-import React, { useState } from "react";
-import { Meteor } from 'meteor/meteor';
-
-export const JoinForm = () => {
-  const [text, setText] = useState("");
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!text) return;
-
-    Meteor.call("joinRoom", { roomId: text }, (error, result) => {
-      if (!error) {
-        console.log("Successfully joined room:", result);
-      }
-    });
-  };
-
-  return (
-    <form className="task-form" onSubmit={handleSubmit}>
-      <input
-        type="text"
-        placeholder="Add Room Code"
-        value={text}
-        onChange={(e) => setText(e.target.value)}
-      />
-      <button type="submit">Submit Room Code</button>
-    </form>
-  );
-};
