--- code/beastly-brawl-showdown/imports/ui/JoinPage.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-import "./JoinPage.css";
-
-export default function JoinPage() {
-  return (
-    <div className="page-container">
-      <h1>Join Page</h1>
-    </div>
-  );
-}
