--- code/beastly-brawl-showdown/imports/ui/Player.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { useParams, useNavigate } from "react-router-dom";
-
-export const Player = () => {
-  const { id } = useParams();
-
-  return (
-    <div>
-      <h1>PLAYER VIEW</h1>
-      <p>Room ID: {id}</p>
-    </div>
-  );
-};
