--- code/beastly-brawl-showdown/imports/ui/Room/HostRoom.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const HostRoom = () => (
-  <>
-    <p>Generating a new room...</p>
-  </>
-);
