+++ code/beastly-brawl-showdown/imports/ui/Room/HostRoomPage.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Meteor } from "meteor/meteor";
 
 export const HostRoomPage = () => {
