--- code/beastly-brawl-showdown/imports/ui/Room/JoinRoom.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-export const JoinRoom = () => {
-  return (
-    <>
-      <h1>Join an existing room</h1>
-      <p>Room Code:</p>
-      <input
-        id="textInput"
-        type="text"
-        // value={inputValue}
-        // onChange={handleChange}
-      />
-    </>
-  );
-};
