+++ code/beastly-brawl-showdown/imports/ui/Room/JoinRoomPage.jsx
@@ -1,4 +1,4 @@
-import { React, useState } from "react";
+import React, { useState } from "react";
 import { Meteor } from "meteor/meteor";
 
 export const JoinRoomPage = () => {
@@ -6,35 +6,35 @@ export const JoinRoomPage = () => {
     <>
       <h1>Join an existing room</h1>
       <p>Room Code:</p>
-      {/* <JoinForm /> */}
+      <JoinForm />
     </>
   );
 };
 
-// export const JoinForm = () => {
-//   const [text, setText] = useState("");
+export const JoinForm = () => {
+  const [text, setText] = useState("");
 
-//   const handleSubmit = async (e) => {
-//     e.preventDefault();
+  const handleSubmit = async (e) => {
+    e.preventDefault();
 
-//     if (!text) return;
+    if (!text) return;
 
-//     Meteor.call("joinRoom", { roomId: text }, (error, result) => {
-//       if (!error) {
-//         console.log("Successfully joined room:", result);
-//       }
-//     });
-//   };
+    Meteor.call("joinRoom", { roomId: text }, (error, result) => {
+      if (!error) {
+        console.log("Successfully joined room:", result);
+      }
+    });
+  };
 
-//   return (
-//     <form className="task-form" onSubmit={handleSubmit}>
-//       <input
-//         type="text"
-//         placeholder="Add Room Code"
-//         value={text}
-//         onChange={(e) => setText(e.target.value)}
-//       />
-//       <button type="submit">Submit Room Code</button>
-//     </form>
-//   );
-// };
+  return (
+    <form className="task-form" onSubmit={handleSubmit}>
+      <input
+        type="text"
+        placeholder="e.g. 123456"
+        value={text}
+        onChange={(e) => setText(e.target.value)}
+      />
+      <button type="submit">Submit Room Code</button>
+    </form>
+  );
+};
